+++ SuperBuild/External_Gadgetron.cmake
@@ -57,6 +57,9 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
   # BLAS
   find_package(BLAS)
   if (APPLE AND NOT (CBLAS_LIBRARY AND CBLAS_INCLUDE_DIR))
+	# if the variables don't exist, let the user set them.
+	SET(CBLAS_LIBRARY "" CACHE FILEPATH "CBAS library")
+	SET(CBLAS_INCLUDE_DIR "" CACHE PATH "CBAS include directory")
     message(FATAL_ERROR "Gadgetron needs CBLAS_LIBRARY and CBLAS_INCLUDE_DIR. If
       these variables do not exist in your CMake, create them manually. CBLAS_LIBRARY
       and CBLAS_INCLUDE_DIR should be FILEPATH and PATH, respectively, and live in
@@ -72,9 +75,9 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
     "Build Gadgetron MATLAB gadgets (not required for SIRF)" ${default_Gadgetron_BUILD_MATLAB_SUPPORT})
   option(Gadgetron_USE_MKL "Instruct Gadgetron to build linking to the MKL. The user must be able to install MKL on his own." OFF)
 
-  option(Gadgetron_USE_CUDA "Enable Gadgetron CUDA (if cuda libraries are present)" ON)
-  mark_as_advanced(Gadgetron_USE_CUDA)
-  
+  option(${proj}_USE_CUDA "Enable ${proj} CUDA (if cuda libraries are present)" ${CUDA_FOUND})
+  mark_as_advanced(${proj}_USE_CUDA)
+
 
   ExternalProject_Add(${proj}
     ${${proj}_EP_ARGS}
@@ -85,28 +88,27 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
     DOWNLOAD_DIR ${${proj}_DOWNLOAD_DIR}
     STAMP_DIR ${${proj}_STAMP_DIR}
     TMP_DIR ${${proj}_TMP_DIR}
-	
-    CMAKE_ARGS
-        -DBUILD_PYTHON_SUPPORT=${Gadgetron_BUILD_PYTHON_SUPPORT}
-        -DBUILD_MATLAB_SUPPORT=${Gadgetron_BUILD_MATLAB_SUPPORT}
-        -DCMAKE_PREFIX_PATH=${SUPERBUILD_INSTALL_DIR}
-        -DCMAKE_LIBRARY_PATH=${SUPERBUILD_INSTALL_DIR}/lib
-        -DCMAKE_INCLUDE_PATH=${SUPERBUILD_INSTALL_DIR}/include
-        -DCMAKE_INSTALL_PREFIX=${Gadgetron_Install_Dir}
-        -DBOOST_INCLUDEDIR=${BOOST_ROOT}/include/
-        -DBOOST_LIBRARYDIR=${BOOST_LIBRARY_DIR}
-        -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE}
-        -DPYTHON_INCLUDE_DIR=${PYTHON_INCLUDE_DIRS}
-        -DPYTHON_LIBRARY=${PYTHON_LIBRARIES}
-        -DGTEST_ROOT=${GTEST_ROOT}
-        ${HDF5_CMAKE_ARGS}
-        -DISMRMRD_DIR=${ISMRMRD_DIR}
-	-DUSE_MKL:BOOL=${Gadgetron_USE_MKL}
-        -DUSE_CUDA=${Gadgetron_USE_CUDA}
-        -DCBLAS_INCLUDE_DIR:PATH=${CBLAS_INCLUDE_DIR}
-        -DCBLAS_LIBRARY:FILEPATH=${CBLAS_LIBRARY}
 
-	    INSTALL_DIR ${Gadgetron_Install_Dir}
+    CMAKE_ARGS
+      -DBUILD_PYTHON_SUPPORT=${Gadgetron_BUILD_PYTHON_SUPPORT}
+      -DBUILD_MATLAB_SUPPORT=${Gadgetron_BUILD_MATLAB_SUPPORT}
+      -DCMAKE_PREFIX_PATH=${SUPERBUILD_INSTALL_DIR}
+      -DCMAKE_LIBRARY_PATH=${SUPERBUILD_INSTALL_DIR}/lib
+      -DCMAKE_INCLUDE_PATH=${SUPERBUILD_INSTALL_DIR}/include
+      -DCMAKE_INSTALL_PREFIX=${Gadgetron_Install_Dir}
+      -DBOOST_INCLUDEDIR=${BOOST_ROOT}/include/
+      -DBOOST_LIBRARYDIR=${BOOST_LIBRARY_DIR}
+      -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE}
+      -DPYTHON_INCLUDE_DIR=${PYTHON_INCLUDE_DIRS}
+      -DPYTHON_LIBRARY=${PYTHON_LIBRARIES}
+      -DGTEST_ROOT=${GTEST_ROOT}
+      ${HDF5_CMAKE_ARGS}
+      -DISMRMRD_DIR=${ISMRMRD_DIR}
+      -DUSE_MKL:BOOL=${Gadgetron_USE_MKL}
+      -DUSE_CUDA=${${proj}_USE_CUDA}
+      -DCBLAS_INCLUDE_DIR:PATH=${CBLAS_INCLUDE_DIR}
+      -DCBLAS_LIBRARY:FILEPATH=${CBLAS_LIBRARY}
+	  INSTALL_DIR ${Gadgetron_Install_Dir}
     DEPENDS
         ${${proj}_DEPENDENCIES}
   )
@@ -120,7 +122,7 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
          COMMAND ${CMAKE_CTEST_COMMAND} -C $<CONFIGURATION> --output-on-failure
          WORKING_DIRECTORY ${${proj}_BINARY_DIR}/test)
   endif()
-     
+
    else()
       if(${USE_SYSTEM_${externalProjName}})
         find_package(${proj} ${${externalProjName}_REQUIRED_VERSION} REQUIRED)